requirements.txt
discord.py>=2.6.0
litellm
httpx
pyyaml
twscrape
youtube-transcript-api
beautifulsoup4
lxml
asyncpraw
Pillow
libsql
tiktoken
requests
pymupdf4llm
